All functionality now complete in C# client.

+++ clients/c-sharp/CachingProxy.cs
@@ -85,7 +85,7 @@ namespace MediaStorage{
 
             Jayrock.Json.JsonObject server = new Jayrock.Json.JsonObject();
             server.Add("host", this.media_server_host);
-            server.Add("port", this.media_server_port);
+            server.Add("port", (int)this.media_server_port);
             proxy.Add("server", server);
             get_json.Add("proxy", proxy);
